+++ src/app/docs/layout/tabs-demo/tabs-dynamic-height/tabs-dynamic-height.component.ts
@@ -1,0 +1,10 @@
+import { Component, ChangeDetectionStrategy } from '@angular/core';
+import { StyleRenderer } from '@alyle/ui';
+
+@Component({
+  selector: 'aui-tabs-dynamic-height',
+  templateUrl: './tabs-dynamic-height.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  providers: [ StyleRenderer ]
+})
+export class TabsDynamicHeightComponent { }