+++ src/lib/carousel/carousel.ts
@@ -149,7 +149,7 @@ export const STYLES = (theme: ThemeVariables & LyCarouselVariables, ref: ThemeRe
         position: relative
         padding: .5em
         outline: none
-      },
+      }
       &>div > span {
         transition: 300ms cubic-bezier(0.65, 0.05, 0.36, 1)
         width: 1em
@@ -159,8 +159,8 @@ export const STYLES = (theme: ThemeVariables & LyCarouselVariables, ref: ThemeRe
         will-change: transform
         display: block
         opacity: .65
-      },
-      '&>div>span.active {
+      }
+      &>div>span.active {
         transform: scale(1)
         opacity: 1
       }